+++ src/config.ts
@@ -34,6 +34,7 @@ export const config = rt
     command: command,
     stackName: rt.String,
     workDir: rt.String,
+    projectName: rt.String,
     commentOnPr: rt.Boolean,
     options: options,
     // Information inferred from the environment that must be present
@@ -59,6 +60,7 @@ export async function makeConfig(): Promise<Config> {
     command: getInput('command', { required: true }),
     stackName: getInput('stack-name', { required: true }),
     workDir: getInput('work-dir') || './',
+    projectName: getInput('project-name') || '',
     secretsProvider: getInput('secrets-provider'),
     cloudUrl: getInput('cloud-url'),
     githubToken: getInput('github-token'),